requirements.txt
streamlit
gradio
tensorflow
numpy
Pillow
gradio_client